Create containers for ui page and divide it into parts
